Run :once fixtures exactly once regardless of test entry point
If the entry point to running a test is `test-var` the :once fixtures
will be run before running the var's :each fixtures.
Nested invocations of `test-var` will not run :once fixtures again

If the entry point is `test-ns` the :once fixtures are run before
running the first tests :each fixtures.